Aha! The type unstable `::Any` in Token was causing the allocs.
The Token struct contains an optional `literal` field, which contains
the _value_ of parsed literal. In Lox, this can only be a Number (parsed
into a Float64) or a (sub)String, extracted from the original string.

Neither of those objects need to be heap allocated. The float is of
course fixed-width, and the substring can be represented as a _view_ on
the original source string that we're parsing from, stored only as a
start and end offset! This is what the `@view(str[a:b])` does.

We were already parsing the string into a substring or a float, but
since we _storing it in a type-unstable field,_ an `::Any`, the value
*needed to be heap-allocated!*

We fix this by manually expanding the `literal::Any` field into a
manually unrolled small union with an _enum_ to represent the type.
Since this object has fixed size (40 bytes), it can be inlined into the
`Token` object. This makes each token 40 bytes bigger, even though most
Tokens don't use it, but avoids the allocation when a Token _does_ use
it, which makes a pretty significant perf improvement!

After:
```julia
julia> str = read("data/test.lox", String)
"// this is a comment\n(( )){} // grouping stuff\n!*+-/=<> <= == // operators\n\nvar language = \"lox\";\n\nprint 2 + 2;"

julia> let str = str^100
       tokens = JuLox.new_tokens(str)
       Profile.clear(); @time Profile.@profile begin
           JuLox.scan_tokens(str, tokens)
       end
       end;
  0.000248 seconds

julia> let str = str^1000
       tokens = JuLox.new_tokens(str)
       Profile.clear(); @time Profile.@profile begin
           JuLox.scan_tokens(str, tokens)
       end
       end;
  0.002332 seconds

julia> using BenchmarkTools

julia> @btime JuLox.scan_tokens($(str), tks) setup=(tks = JuLox.new_tokens(str));
  1.825 μs (0 allocations: 8.74 KiB)
```

For a very small string, the time is _worse_ than before. I think this
is maybe because each token is slightly bigger than before and/or takes
longer to initialize?

However, for very long strings, the time is quite reduced, presumably
because of the reduced allocation time and the reduced GC time:
```julia
julia> let str = str^10000
       tokens = JuLox.new_tokens(str)
       Profile.clear(); @time Profile.@profile begin
           JuLox.scan_tokens(str, tokens)
       end
       end;
  0.021979 seconds

julia> let str = str^100000
       tokens = JuLox.new_tokens(str)
       Profile.clear(); @time Profile.@profile begin
           JuLox.scan_tokens(str, tokens)
       end
       end;
  0.169508 seconds

julia> let str = str^1000000
       tokens = JuLox.new_tokens(str)
       Profile.clear(); @time Profile.@profile begin
           JuLox.scan_tokens(str, tokens)
       end
       end;
  1.764036 seconds
```

There are some small remaining gains to be had from:
- `@inbounds` to remove bounds-checking
- better preallocation of the tokens vector
- indexing via codeunits, not char indexes
- replacing the KEYWORDS dict with another switch statement
but this is already looking pretty good. :)